requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.5.2=py310haa95532_0
beautifulsoup4=4.12.2=pyha770c72_0
blinker=1.5=pyhd8ed1ab_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotli-python=1.0.9=py310h8a704f9_7
brotlicffi=1.0.9.2=py310h8a704f9_2
brotlipy=0.7.0=py310he2412df_1004
bzip2=1.0.8=he774522_0
ca-certificates=2022.12.7=h5b45459_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310hcbf9ad4_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
crispy-bootstrap5=0.6=py_0
cryptography=3.4.8=py310ha857299_1
defusedxml=0.7.1=pyhd8ed1ab_0
django=4.1=py310haa95532_0
django-allauth=0.53.1=pyhd8ed1ab_0
django-bootstrap5=23.1=pyhd8ed1ab_0
django-crispy-forms=2.0=pyhd8ed1ab_0
freetype=2.12.1=ha860e81_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=2.0.0=py_1
jpeg=9e=h2bbff1b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libpng=1.6.37=h2a8f88b_0
libtiff=4.5.0=h6c2663c_1
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
mutagen=1.46.0=pyhd8ed1ab_0
oauthlib=3.2.2=pyhd8ed1ab_0
openssl=1.1.1t=h2bbff1b_0
pillow=9.3.0=py310hd77b12b_2
pip=23.0.1=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pycryptodomex=3.15.0=py310h2bbff1b_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
python=3.10.9=h966fe2a_0
python-tzdata=2021.1=pyhd3eb1b0_0
python3-openid=3.2.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytube=12.1.3=pyhd8ed1ab_0
requests=2.28.2=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
setuptools=65.6.3=py310haa95532_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.40.1=h2bbff1b_0
sqlparse=0.4.3=py310haa95532_0
tk=8.6.12=h2bbff1b_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.15=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
websockets=10.4=py310h2bbff1b_1
wheel=0.38.4=py310haa95532_0
whitenoise=6.4.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wincertstore=0.2=py310haa95532_2
xz=5.2.10=h8cc25b3_1
youtube-dl=2021.12.17=pyhd8ed1ab_0
yt-dlp=2023.3.4=pyhd8ed1ab_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
